Fix UID reporting properly
The patch that was already committed apparently fixes the issue but
using a non-documented method. This patch fixes the issue as described in
"man proc". (The UID is extracted by parsing /proc/*/status not by using
stat() on /proc/*/task). Also, this patch can easily be extended to also
report other UIDs besides the "effective" UID: real, effective, saved
set, and filesystem UIDs.